physa-ci-app/requirements.txt
azure-functions
azure-storage-queue
azure-storage-file
azure-cosmosdb-table
pyjwt
requests